fix(sys_call): move system call to a separate mod
