+++ os/src/main.rs
@@ -2,26 +2,10 @@
 #![no_main]
 
 mod lang_item;
+mod sys;
 
-const SYSCALL_EXIT: usize = 93;
-
-fn syscall(id: usize, args: [usize; 3]) -> isize {
-    let mut ret;
-    unsafe {
-        core::arch::asm!(
-        "ecall",
-        inlateout("x10") args[0] => ret,
-        in("x11") args[1],
-        in("x12") args[2],
-        in("x17") id,
-        );
-    }
-    ret
-}
-
-pub fn sys_exit(xstate: i32) -> isize {
-    syscall(SYSCALL_EXIT, [xstate as usize, 0, 0])
-}
+use crate::console::*;
+use crate::sys::*;
 
 #[no_mangle]
 extern "C" fn _start() {
